feat: add download as PDF option
* There's three ways to get it.

1. Either take a screenshot
of the node (e.g. html2canva or similar) which doesn't
seem to work with tailwind & shadcn's oklch colors and
cause a heavier PDF and with non-selectable text.
2. Another alternative is to use a different package
that can output a vectorized PDF. However, the downside
is that styles have to be passed when creating it,
which is painful given we support templates too.
3. Use browser's print API, which I chose.